Fix up warnings

+++ Source/Classes/PINRemoteImageDownloadTask.h
@@ -17,7 +17,7 @@
 @property (nonatomic, assign) CFTimeInterval sessionTaskEndTime;
 @property (nonatomic, assign) BOOL hasProgressBlocks;
 @property (nonatomic, strong, nullable) NSString *ifRange;
-@property (nonatomic, assign) NSUInteger totalBytes;
+@property (nonatomic, assign) long long totalBytes;
 @property (nonatomic, strong, nullable) PINResume *resume;
 @property (nonatomic, strong, nullable) PINProgressiveImage *progressImage;
