+++ Source/Classes/PINResume.h
@@ -13,11 +13,11 @@ NS_ASSUME_NONNULL_BEGIN
 @interface PINResume : NSObject
 
 - (id)init NS_UNAVAILABLE;
-+ (PINResume *)resumeData:(NSData *)resumeData ifRange:(NSString *)ifRange totalBytes:(NSUInteger)totalBytes;
++ (PINResume *)resumeData:(NSData *)resumeData ifRange:(NSString *)ifRange totalBytes:(long long)totalBytes;
 
 @property (nonatomic, strong, readonly) NSData *resumeData;
 @property (nonatomic, strong, readonly) NSString *ifRange;
-@property (nonatomic, assign, readonly) NSUInteger totalBytes;
+@property (nonatomic, assign, readonly) long long totalBytes;
 
 @end
